delete IOParser.kt
